refactored favourites page
